model/requirements.txt
google-generativeai
fastapi
uvicorn
PyPDF2
zipfile36
